@@ -1,23 +1,42 @@
 module github.com/Dreamacro/clash
 
-go 1.14
+go 1.19
 
 require (
-	github.com/Dreamacro/go-shadowsocks2 v0.1.6-0.20200722122336-8e5c7db4f96a
-	github.com/eapache/queue v1.1.0 // indirect
-	github.com/go-chi/chi v4.1.2+incompatible
-	github.com/go-chi/cors v1.1.1
-	github.com/go-chi/render v1.0.1
-	github.com/gofrs/uuid v3.3.0+incompatible
-	github.com/gorilla/websocket v1.4.2
-	github.com/miekg/dns v1.1.29
-	github.com/oschwald/geoip2-golang v1.4.0
-	github.com/sirupsen/logrus v1.6.0
-	github.com/stretchr/testify v1.6.1
-	golang.org/x/crypto v0.0.0-20200604202706-70a84ac30bf9
-	golang.org/x/net v0.0.0-20200602114024-627f9648deb9
-	golang.org/x/sync v0.0.0-20200317015054-43a5402ce75a
-	golang.org/x/sys v0.0.0-20200323222414-85ca7c5b95cd
-	gopkg.in/eapache/channels.v1 v1.1.0
-	gopkg.in/yaml.v2 v2.3.0
+	github.com/go-chi/chi/v5 v5.0.8
+	github.com/go-chi/cors v1.2.1
+	github.com/go-chi/render v1.0.2
+	github.com/gofrs/uuid v4.4.0+incompatible
+	github.com/gorilla/websocket v1.5.0
+	github.com/insomniacslk/dhcp v0.0.0-20221215072855-de60144f33f8
+	github.com/mdlayher/netlink v1.7.2-0.20221213171556-9881fafed8c7
+	github.com/miekg/dns v1.1.50
+	github.com/oschwald/geoip2-golang v1.8.0
+	github.com/samber/lo v1.37.0
+	github.com/sirupsen/logrus v1.9.0
+	github.com/stretchr/testify v1.8.1
+	go.etcd.io/bbolt v1.3.6
+	go.uber.org/atomic v1.10.0
+	go.uber.org/automaxprocs v1.5.1
+	golang.org/x/crypto v0.5.0
+	golang.org/x/net v0.5.0
+	golang.org/x/sync v0.1.0
+	golang.org/x/sys v0.4.0
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/ajg/form v1.5.1 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/go-cmp v0.5.9 // indirect
+	github.com/josharian/native v1.1.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/mdlayher/socket v0.4.0 // indirect
+	github.com/oschwald/maxminddb-golang v1.10.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/u-root/uio v0.0.0-20221213070652-c3537552635f // indirect
+	golang.org/x/exp v0.0.0-20220303212507-bbda1eaf7a17 // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
+	golang.org/x/text v0.6.0 // indirect
+	golang.org/x/tools v0.1.12 // indirect
 )
