@@ -1,9 +1,11 @@
 package dns
 
 import (
+	"errors"
 	"net"
 
 	"github.com/Dreamacro/clash/common/sockopt"
+	"github.com/Dreamacro/clash/context"
 	"github.com/Dreamacro/clash/log"
 
 	D "github.com/miekg/dns"
@@ -21,24 +23,35 @@ type Server struct {
 	handler handler
 }
 
+// ServeDNS implement D.Handler ServeDNS
 func (s *Server) ServeDNS(w D.ResponseWriter, r *D.Msg) {
-	if len(r.Question) == 0 {
+	msg, err := handlerWithContext(s.handler, r)
+	if err != nil {
 		D.HandleFailed(w, r)
 		return
 	}
+	msg.Compress = true
+	w.WriteMsg(msg)
+}
 
-	s.handler(w, r)
+func handlerWithContext(handler handler, msg *D.Msg) (*D.Msg, error) {
+	if len(msg.Question) == 0 {
+		return nil, errors.New("at least one question is required")
+	}
+
+	ctx := context.NewDNSContext(msg)
+	return handler(ctx, msg)
 }
 
 func (s *Server) setHandler(handler handler) {
 	s.handler = handler
 }
 
-func ReCreateServer(addr string, resolver *Resolver) error {
+func ReCreateServer(addr string, resolver *Resolver, mapper *ResolverEnhancer) {
 	if addr == address && resolver != nil {
-		handler := newHandler(resolver)
+		handler := newHandler(resolver, mapper)
 		server.setHandler(handler)
-		return nil
+		return
 	}
 
 	if server.Server != nil {
@@ -47,33 +60,47 @@ func ReCreateServer(addr string, resolver *Resolver) error {
 		address = ""
 	}
 
+	if addr == "" {
+		return
+	}
+
+	var err error
+	defer func() {
+		if err != nil {
+			log.Errorln("Start DNS server error: %s", err.Error())
+		}
+	}()
+
 	_, port, err := net.SplitHostPort(addr)
 	if port == "0" || port == "" || err != nil {
-		return nil
+		return
 	}
 
 	udpAddr, err := net.ResolveUDPAddr("udp", addr)
 	if err != nil {
-		return err
+		return
 	}
 
 	p, err := net.ListenUDP("udp", udpAddr)
 	if err != nil {
-		return err
+		return
 	}
 
 	err = sockopt.UDPReuseaddr(p)
 	if err != nil {
 		log.Warnln("Failed to Reuse UDP Address: %s", err)
+
+		err = nil
 	}
 
 	address = addr
-	handler := newHandler(resolver)
+	handler := newHandler(resolver, mapper)
 	server = &Server{handler: handler}
 	server.Server = &D.Server{Addr: addr, PacketConn: p, Handler: server}
 
 	go func() {
 		server.ActivateAndServe()
 	}()
-	return nil
+
+	log.Infoln("DNS server listening at: %s", p.LocalAddr().String())
 }
