@@ -4,13 +4,15 @@ import (
 	"net/http"
 	"path/filepath"
 
+	"github.com/Dreamacro/clash/component/resolver"
 	"github.com/Dreamacro/clash/config"
+	"github.com/Dreamacro/clash/constant"
 	"github.com/Dreamacro/clash/hub/executor"
+	P "github.com/Dreamacro/clash/listener"
 	"github.com/Dreamacro/clash/log"
-	P "github.com/Dreamacro/clash/proxy"
 	"github.com/Dreamacro/clash/tunnel"
 
-	"github.com/go-chi/chi"
+	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/render"
 )
 
@@ -20,17 +22,6 @@ func configRouter() http.Handler {
 	r.Put("/", updateConfigs)
 	r.Patch("/", patchConfigs)
 	return r
-}
-
-type configSchema struct {
-	Port        *int               `json:"port"`
-	SocksPort   *int               `json:"socks-port"`
-	RedirPort   *int               `json:"redir-port"`
-	MixedPort   *int               `json:"mixed-port"`
-	AllowLan    *bool              `json:"allow-lan"`
-	BindAddress *string            `json:"bind-address"`
-	Mode        *tunnel.TunnelMode `json:"mode"`
-	LogLevel    *log.LogLevel      `json:"log-level"`
 }
 
 func getConfigs(w http.ResponseWriter, r *http.Request) {
@@ -47,8 +38,19 @@ func pointerOrDefault(p *int, def int) int {
 }
 
 func patchConfigs(w http.ResponseWriter, r *http.Request) {
-	general := &configSchema{}
-	if err := render.DecodeJSON(r.Body, general); err != nil {
+	general := struct {
+		Port        *int               `json:"port"`
+		SocksPort   *int               `json:"socks-port"`
+		RedirPort   *int               `json:"redir-port"`
+		TProxyPort  *int               `json:"tproxy-port"`
+		MixedPort   *int               `json:"mixed-port"`
+		AllowLan    *bool              `json:"allow-lan"`
+		BindAddress *string            `json:"bind-address"`
+		Mode        *tunnel.TunnelMode `json:"mode"`
+		LogLevel    *log.LogLevel      `json:"log-level"`
+		IPv6        *bool              `json:"ipv6"`
+	}{}
+	if err := render.DecodeJSON(r.Body, &general); err != nil {
 		render.Status(r, http.StatusBadRequest)
 		render.JSON(w, r, ErrBadRequest)
 		return
@@ -63,10 +65,15 @@ func patchConfigs(w http.ResponseWriter, r *http.Request) {
 	}
 
 	ports := P.GetPorts()
-	P.ReCreateHTTP(pointerOrDefault(general.Port, ports.Port))
-	P.ReCreateSocks(pointerOrDefault(general.SocksPort, ports.SocksPort))
-	P.ReCreateRedir(pointerOrDefault(general.RedirPort, ports.RedirPort))
-	P.ReCreateMixed(pointerOrDefault(general.MixedPort, ports.MixedPort))
+
+	tcpIn := tunnel.TCPIn()
+	udpIn := tunnel.UDPIn()
+
+	P.ReCreateHTTP(pointerOrDefault(general.Port, ports.Port), tcpIn)
+	P.ReCreateSocks(pointerOrDefault(general.SocksPort, ports.SocksPort), tcpIn, udpIn)
+	P.ReCreateRedir(pointerOrDefault(general.RedirPort, ports.RedirPort), tcpIn, udpIn)
+	P.ReCreateTProxy(pointerOrDefault(general.TProxyPort, ports.TProxyPort), tcpIn, udpIn)
+	P.ReCreateMixed(pointerOrDefault(general.MixedPort, ports.MixedPort), tcpIn, udpIn)
 
 	if general.Mode != nil {
 		tunnel.SetMode(*general.Mode)
@@ -76,16 +83,18 @@ func patchConfigs(w http.ResponseWriter, r *http.Request) {
 		log.SetLevel(*general.LogLevel)
 	}
 
+	if general.IPv6 != nil {
+		resolver.DisableIPv6 = !*general.IPv6
+	}
+
 	render.NoContent(w, r)
 }
 
-type updateConfigRequest struct {
-	Path    string `json:"path"`
-	Payload string `json:"payload"`
-}
-
 func updateConfigs(w http.ResponseWriter, r *http.Request) {
-	req := updateConfigRequest{}
+	req := struct {
+		Path    string `json:"path"`
+		Payload string `json:"payload"`
+	}{}
 	if err := render.DecodeJSON(r.Body, &req); err != nil {
 		render.Status(r, http.StatusBadRequest)
 		render.JSON(w, r, ErrBadRequest)
@@ -104,9 +113,12 @@ func updateConfigs(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 	} else {
+		if req.Path == "" {
+			req.Path = constant.Path.Config()
+		}
 		if !filepath.IsAbs(req.Path) {
 			render.Status(r, http.StatusBadRequest)
-			render.JSON(w, r, newError("path is not a absoluted path"))
+			render.JSON(w, r, newError("path is not a absolute path"))
 			return
 		}
 
